Merge crossbar config with OCS crossbar config
Our goal is for a sisock DataNodeServer to subscribe to data feeds provided by
OCS to provide live monitoring in grafana. While the autobahn documentation
leads one to believe a single Component can communicate with multiple crossbar
servers, in practice we have yet to implement that. This led us to using a
common crossbar server.

One change here is the realm name. We've defaulted both sisock and OCS to use a
realm called 'test_realm'. This provides the common namespace for pub/sub and
RPC for use in development environments.

For compatability with OCS we've added the OCS roles, 'iocs_agent' and
'iocs_controller', and the unsecured websocket transport on port 8001.

For interoperability between sisock and OCS we've also switched the auth on
port 8001 to use the 'server' role. This is needed, as sisock needs to
communicate with uri's that don't match the 'observatory.' prefix enforced for
the 'iocs_agent' role. If this uri prefix for OCS needs to be this strict this
should be fixed, but at the time of writing it's not entirely clear to the
author the best way to accomplish this and maintain sisock/ocs communication.

This config allows sisock and ocs to run off of the same crossbar server, as
well as communicate with each other.