Have _search rebuild node::field list when called
If a field is added after the data node as been added to the sisock hub then
the field won't be discovered. _search is called whenever a user views the drop
down menu of timeseries in grafana, and so if a new field as since been added
to a data node, it will show up in the drop down.