Use Docker Compose to build and run the containers
As the number of DataNodeServers grows, so does the number of Docker
containers. Docker Compose is a tool for defining and running multi-container
applications. This commit configures docker-compose and addresses some runtime
connection issues that arise from trying to start all the servers at once.

Ideally each container's app is able to survive the disappearance of other
containers, perhaps those they depend on. One issue addressed here is the
grafana http json server trying to make an RPC call prior to the hub
registering the call with the crossbar server, but after the crossbar server is
accepting connections. The result was a dead grafana http json server. We now
handle the exception and wait a few seconds for the hub to finish starting up,
then try again.

Similarly, the DataNodeServers could run into the same issue. I've solved this
instead by having them wait a few seconds few seconds before starting, though a
similar retry mechanism could be put in place in the DataNodeSever base class.