Add max returned data option to radiometer server
When getting data, we were reading everything from disk and sending it all. So
larger datasets took more bandwidth and so took longer to get to the user. Now
we can set a MAX_POINTS environment variable when starting the container that
limits the maximum number of data points sent in response to a get_data call.