Expose crossbar ports only to localhost
Docker manipulates iptables for its own networking purposes. This has the side
effect of exposing ports to all IPs even if something like ufw is configured
to drop those connections. If we explicitly define the IP to which we expose
the ports we avoid this behavior. This means in this configuration only OCS
Agents running on the same computer as the sisock stack will be able to
connect. This is fine until we run in a configuration where OCS Agents are
distributed among multiple computers.

We also remove the port 5000 exposure to sisock_grafana_http as this is only
needed internally on the sisock network.