Use Dockerfile best practices in prep for registry
In preparation for using the Docker Hub registry I explored pushing to a
private registry to see what sort of implications that had for the build
process. This commit updates all the Dockerfiles in the repo to explicitly use
the latest build of sisock. I've also rearranged so that the DataNodeServers
aren't based on the crossbar image, but instead the base sisock image. The
expectation now is that you bind mount the TLS certs into the container (which
was previously covered by the fact that they were copied into the crossbar
container, not a great practice in itself.)