Update add_files_to_feeds_table for new so3g HKAgg format
This also changes the table format a bit, since previously the feed was in each
Frame. Now what was the feed is the 'description' for a provider, and each
frame has the corresponding provider ID. We add both the prov_id and
description to the feeds table and use the prov_id like we used to use the feed
name itself.

The feeds table format now looks like this:

```
MariaDB [files]> describe feeds;
+-------------+--------------+------+-----+---------+-------+
| Field       | Type         | Null | Key | Default | Extra |
+-------------+--------------+------+-----+---------+-------+
| id          | int(11)      | NO   | PRI | NULL    |       |
| filename    | varchar(255) | YES  | MUL | NULL    |       |
| path        | varchar(255) | YES  |     | NULL    |       |
| prov_id     | int(11)      | YES  |     | NULL    |       |
| description | varchar(255) | YES  |     | NULL    |       |
| scanned     | tinyint(1)   | NO   |     | 0       |       |
+-------------+--------------+------+-----+---------+-------+
6 rows in set (0.002 sec)
```