Prevent caching of incomplete data files
This fixes a bug that would crop up when we tried opening a data file that was
still being written. The cache would think the file has been opened, and thus
not open it again to avoid excess I/O. This resulted in gaps in the data
depending on how often you tried querying for the data.