Accept new OCS feed structure in thermometry server
The feed message structure changed after OCS v0.4.2. This patch updates the way
the thermometry data node server parses this feed message and is first
compatible with OCS v0.4.3.

We also take this time to add a buffer duration argument which can be set in
the environment variables for a thermometry data node server container using
the variable "BUFFER_TIME".