Move construction of field list to g3-file-scanner
The loop which builds the list of fields which the g3-reader would return when
get_fields() was called queries the SQL database an excessively. get_fields()
is called with each grafana search query to the HTTP server. This occurs once
for each drop down box present in the panel configurations "metrics" tab. Given
a large dataset and many metrics, this makes the g3-reader egregiously slow.

This moves the construction of this field list over to the g3-file-scanner,
which only is expected to scan on hour long timescales, so it's alright if the
build takes a couple of seconds. g3-reader now gets this field list by grabbing
everything in the new 'description' table's singular column. This dropped query
times by a factor of 1000.