Automatically upgrade database structure to version 1
Moving from version 0 of the database table structure to version 1 we've moved
the filename and path information to a separate table, where we gather other
information about the files such as their size and availability. This will
allow us to easily keep track of the amount of data open as well as the
availability of data in the event it is modified or moved.