Handle processing for already configured fields
The exception catching a KeyError wasn't working for already configured fields
recently. This imeplements the same work around as we were doing for the
g3-reader. For fields that we can't find we just return an empty array, so if
someone has already configured a query in Grafana we'll give them nothing back.
This results in the field name still showing in the key of their plot (so they
won't be confused why their query isn't showing up in the plot), but no data
being displayed. Allows fields also configured for the same plot that do have
data to be plotted.